Clean up the System.output inside the tests.
Inside tests, we should never depend on System.output to catch the failure.
Several tests are not checked by Assert, but only by System.output (not checked at all).
This is the first step to make the tests log more clear.
The next step is digging into each test and cover all the checks into Assert.